ci-scripts: packaging: add control variable WITHOUT_PACKAGING
In Travis CI, if a build time is long, build stops with an error
exceeding max time on the job.  If you want to test with clang, it must
run in parallel environment.
However, deploy tasks of gcc and clang environment will upload the same
package, so need to suppress one side.

If environments that WITHOUT_PACKAGING=1 is set, it will not upload the
package.  This variable is defined in environment with clang test.