ci-scripts: packaging: add control variable SIMULATE_UPLOAD
If SIMULATE_UPLOAD=1 is set, add "-s" option to dput command in deploy
task to simulate uploading only.  DEBUG variable only switches display
of debug messages and does not change the upload command.

This SIMULATE_UPLOAD variable should be specified in CI services
interface, not config file of packaging script, so that behavior of
script can be changes without commit/push to repository.