Invalidate Link::Rs
The Link class implementation used the coordinate conversion where each link's local offset rotation from its parent link was cleared to increse the performance of the forward kinematics calculation. In this conversion, the global rotation matrix of each link was different from the value based on the original model description. This design required programmers to implement the additional coordinate conversions between the attitudes based on the original model and the attitudes for the internal calculation, and it caused a confusion in programming using the Link class. To resolve this problem, the implementation was modified to use the original local offset rotation from the parent link, and the member variable Rs, which was used for the attitude conversion, was invalidated. Now Link::T() and its related members such as Link::R() contain the atttitude based on the original model and no coordinate conversion is required for those values.