Add BodySelectionManager::setItemTreeWidgetToSelectCurrentBodyItem
When an item tree widget is specified by this function, calling BodySelectionManager::setCurrent with doSelectBodyItem true selects the item as if the item were selected by a user on the specified item tree widget. If the item tree widget belongs to a selection sync group, the item is exclusively selected within the group.